Only check for standby flag, if it is configured
git-svn-id: svn://85.10.198.214/trunk@178 398de0b3-2755-4f39-ad97-b6641548fef7

+++ EnhancedMovieCenter.py
@@ -111,10 +111,11 @@ def EMCStartup(session):
 	DelayedFunction(5000, cleanupSetup)
 
 	# Go into standby if the reason for restart was EMC auto-restart
-	if os.path.exists(config.EMC.folder.value + "/EMC_standby_flag.tmp"):
-		emcDebugOut("+++ Going into Standby mode after auto-restart")
-		Notifications.AddNotification(Screens.Standby.Standby)
-		emcTasker.shellExecute("rm -f " + config.EMC.folder.value + "/EMC_standby_flag.tmp")
+	if config.EMC.enigmarestart.value:
+		if os.path.exists(config.EMC.folder.value + "/EMC_standby_flag.tmp"):
+			emcDebugOut("+++ Going into Standby mode after auto-restart")
+			Notifications.AddNotification(Screens.Standby.Standby)
+			emcTasker.shellExecute("rm -f " + config.EMC.folder.value + "/EMC_standby_flag.tmp")
 
 class EnhancedMovieCenterMenu(ConfigListScreen, Screen):
 	skin = """
